moved all types to new file reduced includes everywhere
