+++ main/board.h
@@ -1,15 +1,8 @@
 #include "constants.h"
-#include "snake.h"
+#include "types.h"
 
 #ifndef BOARD_H
 #define BOARD_H
-
-
-// typedef struct {
-//     short int piece_type;
-//     body_t* part;
-// } board_piece_t;
-
 
 void init_board(board_piece_t* board);
 board_piece_t get_square_value(board_piece_t* board, int x, int y);
@@ -19,6 +12,5 @@ void place_random_food(board_piece_t* board, int* count_food);
 void place_food_at_coords(board_piece_t* board, int x, int y, int* count_food);
 char convert_board_int(board_piece_t piece);
 void print_board(board_piece_t* board);
-
 
 #endif
